requirements.txt
tweepy
pillow
urllib
